+++ docs/release_notes.rst
@@ -263,7 +263,7 @@ Data Coverage
   The newly accessible tables include:
 
   * :ref:`core_ferc714__respondent_id` (linking FERC-714 respondents to EIA utilities)
-  * :ref:`core_ferc714__hourly_demand_by_planning_area` (hourly electricity demand by
+  * :ref:`out_ferc714__hourly_planning_area_demand` (hourly electricity demand by
     planning area)
   * :ref:`out_ferc714__respondents_with_fips` (annual respondents with county FIPS IDs)
   * :ref:`out_ferc714__summarized_demand` (annual demand for FERC-714 respondents)